+++ jre_emul/android/libcore/luni/src/main/native/java_lang_StringToReal.m
@@ -212,7 +212,7 @@ static jdouble createDouble(const char* s, jint e) {
    *   3. (unprocessed digits + e) > 0 indicates that the value is
    *      simply too big to be stored as a double, so return Infinity
    */
-  if ((unprocessedDigits = strlen (s)) > 0)
+  if ((unprocessedDigits = (int) strlen (s)) > 0)
     {
       e += unprocessedDigits;
       if (index > -1)
@@ -651,7 +651,7 @@ static jfloat createFloat(const char* s, jint e) {
    *   3. (unprocessed digits + e) > 0 indicates that the value is
    *      simply too big to be stored as a double, so return Infinity
    */
-  if ((unprocessedDigits = strlen (s)) > 0)
+  if ((unprocessedDigits = (int) strlen (s)) > 0)
     {
       e += unprocessedDigits;
       if (index > -1)
